Avoid re-resolving methods from tokens unless necessary
Make the condition under which we track a reference to a method from shared code as a `RuntimeDeterminedMethod` (as opposed to a simple entrypoint) a bit looser.

The only circumstance under which we need to know the exact reference is e.g.

```csharp
class Program
{
  static bool Do2<T>(object x)
  {
    return x is T[,,,];
  }

  static bool Do1<T>(object x)
  {
    return Do2<T>(x);
  }

  static void Main()
  {
    Do1<object>(new Object());
  }
}
```

In the above example, we need to track the reference from Do1 to Do2 as a RuntimeDetermined dependency so that the shadow method for Do1<object> ends up depending on the shadow method for Do2<object>. This is needed because the generic dictionary for Do2<object> doesn't depend on the shadow method for Do2<object> (i.e. without the tracking, we would end up knowing we need the dictionary, but nothing would make sure we can actually populate the entries).

The other potential fix would be to make method generic dictionaries depend on the shadow method they represent.

The reason why we need to be more careful about this is devirtualization in RyuJIT - if we devirtualized a method, the pResolvedToken RyuJIT is passing us is bogus and can't be used to get a RuntimeDetermined method.